staslib: break stas.py into smaller modules
Added 2 new modules:
- ctrl.py
- service.py
