+++ staslib/meson.build
@@ -21,9 +21,11 @@ files_to_copy = [
     '__init__.py',
     'avahi.py',
     'conf.py',
+    'ctrl.py',
     'glibudev.py',
     'gutil.py',
     'log.py',
+    'service.py',
     'stas.py',
     'trid.py',
     'udev.py',
