+++ test/test-controller.py
@@ -1,9 +1,10 @@
 #!/usr/bin/python3
 import os
 import unittest
-from staslib import stas, log, conf, trid
+from staslib import log, conf, ctrl, trid
 from libnvme import nvme
 from pyfakefs.fake_filesystem_unittest import TestCase
+from gi.repository import GLib
 
 class Test(TestCase):
     '''Unit tests for class Controller'''
@@ -26,33 +27,37 @@ class Test(TestCase):
             'host-iface':  'wlp0s20f3',
         }
 
+        self._loop = GLib.MainLoop()
+        self.controller = ctrl.Controller(root=self.NVME_ROOT, host=self.NVME_HOST, tid=trid.TID(self.NVME_TID))
+
+    def tearDown(self):
+        self._loop.quit()
+
     def test_get_device(self):
-        ctrl = stas.Controller(root=self.NVME_ROOT, host=self.NVME_HOST, tid=trid.TID(self.NVME_TID))
-        self.assertEqual(ctrl._connect_attempts, 0)
-        self.assertRaises(NotImplementedError, ctrl._try_to_connect)
-        self.assertEqual(ctrl._connect_attempts, 1)
-        self.assertRaises(NotImplementedError, ctrl._find_existing_connection)
-        self.assertEqual(ctrl.id, "(tcp, 10.10.10.10, 8009, nqn.1988-11.com.dell:SFSS:2:20220208134025e8, wlp0s20f3, 1.2.3.4)")
-        # raise Exception(ctrl._connect_op)
-        self.assertEqual(str(ctrl.tid), "(tcp, 10.10.10.10, 8009, nqn.1988-11.com.dell:SFSS:2:20220208134025e8, wlp0s20f3, 1.2.3.4)")
-        self.assertEqual(ctrl.device, '')
-        self.assertEqual(str(ctrl.controller_id_dict()), "{'transport': 'tcp', 'traddr': '10.10.10.10', 'trsvcid': '8009', 'host-traddr': '1.2.3.4', 'host-iface': 'wlp0s20f3', 'subsysnqn': 'nqn.1988-11.com.dell:SFSS:2:20220208134025e8', 'device': ''}")
-        # self.assertEqual(ctrl.details(), "{'transport': 'tcp', 'traddr': '10.10.10.[265 chars]ff]'}")
-        self.assertEqual(ctrl.info(), {'transport': 'tcp', 'traddr': '10.10.10.10', 'trsvcid': '8009', 'host-traddr': '1.2.3.4', 'host-iface': 'wlp0s20f3', 'subsysnqn': 'nqn.1988-11.com.dell:SFSS:2:20220208134025e8', 'device': '', 'hostid': '', 'hostnqn': '', 'model': '', 'serial': '', 'connect attempts': '1', 'retry connect timer': '60.0s [off]'})
-        # print(ctrl._connect_op)
-        self.assertEqual(ctrl.cancel(), None)
-        self.assertEqual(ctrl.kill(), None)
-        # self.assertEqual(ctrl.disconnect(), 0)
+        self.assertEqual(self.controller._connect_attempts, 0)
+        self.assertRaises(NotImplementedError, self.controller._try_to_connect)
+        self.assertEqual(self.controller._connect_attempts, 1)
+        self.assertRaises(NotImplementedError, self.controller._find_existing_connection)
+        self.assertEqual(self.controller.id, "(tcp, 10.10.10.10, 8009, nqn.1988-11.com.dell:SFSS:2:20220208134025e8, wlp0s20f3, 1.2.3.4)")
+        # raise Exception(self.controller._connect_op)
+        self.assertEqual(str(self.controller.tid), "(tcp, 10.10.10.10, 8009, nqn.1988-11.com.dell:SFSS:2:20220208134025e8, wlp0s20f3, 1.2.3.4)")
+        self.assertEqual(self.controller.device, '')
+        self.assertEqual(str(self.controller.controller_id_dict()), "{'transport': 'tcp', 'traddr': '10.10.10.10', 'trsvcid': '8009', 'host-traddr': '1.2.3.4', 'host-iface': 'wlp0s20f3', 'subsysnqn': 'nqn.1988-11.com.dell:SFSS:2:20220208134025e8', 'device': ''}")
+        # self.assertEqual(self.controller.details(), "{'transport': 'tcp', 'traddr': '10.10.10.[265 chars]ff]'}")
+        self.assertEqual(self.controller.info(), {'transport': 'tcp', 'traddr': '10.10.10.10', 'trsvcid': '8009', 'host-traddr': '1.2.3.4', 'host-iface': 'wlp0s20f3', 'subsysnqn': 'nqn.1988-11.com.dell:SFSS:2:20220208134025e8', 'device': '', 'hostid': '', 'hostnqn': '', 'model': '', 'serial': '', 'connect attempts': '1', 'retry connect timer': '60.0s [off]'})
+        # print(self.controller._connect_op)
+        self.assertEqual(self.controller.cancel(), None)
+        self.assertEqual(self.controller.kill(), None)
+        # self.assertEqual(self.controller.disconnect(), 0)
 
     def test_connect(self):
-        ctrl = stas.Controller(root=self.NVME_ROOT, host=self.NVME_HOST, tid=trid.TID(self.NVME_TID))
-        self.assertEqual(ctrl._connect_attempts, 0)
-        ctrl._find_existing_connection = lambda : None
+        self.assertEqual(self.controller._connect_attempts, 0)
+        self.controller._find_existing_connection = lambda : None
         with self.assertLogs(logger=log.LOG, level='DEBUG') as captured:
-            ctrl._try_to_connect()
+            self.controller._try_to_connect()
         self.assertEqual(len(captured.records), 1)
         self.assertEqual(captured.records[0].getMessage(), "Controller._try_to_connect()       - (tcp, 10.10.10.10, 8009, nqn.1988-11.com.dell:SFSS:2:20220208134025e8, wlp0s20f3, 1.2.3.4) Connecting to nvme control with cfg={'hdr_digest': False, 'data_digest': False}")
-        self.assertEqual(ctrl._connect_attempts, 1)
+        self.assertEqual(self.controller._connect_attempts, 1)
 
 
 if __name__ == '__main__':
